Unloaded Event Removed, added uc cashing, added loader and async init, dal class complete

+++ Models/clsImagePhotoFlipper.cs
@@ -1,6 +1,7 @@
 ﻿using GroepswerkTaak1.Models.Base;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,9 +10,9 @@ namespace GroepswerkTaak1.Models
 {
 	public class clsImagePhotoFlipper: clsBaseModel
 	{		
-		public int ImagePhotoFlipperID { get; set; }
+		public short ImagePhotoFlipperID { get; set; }
 		public required byte[] ImageBytes {  get; set; }
-		public required object ControlField { get; set; }
-		
+		public required object ControlField { get; set; }	
+
 	}
 }
